- improved covertToMap method - solved duration problem. - added versions property to song class - authenticate user before any page will be loaded.

+++ dart_angular/lib/src/class/archive/playlist.dart
@@ -60,7 +60,7 @@ class Playlist implements SongItem
 
   String getDuration()
   {
-    int length = 0;
+    double length = 0;
     list.forEach((media){ length += media.duration; });
 
     int hours = (length / 3600).floor();
